Move BTPurchaseUnit
